+++ vector/src/main/java/im/vector/app/features/autocomplete/command/AutocompleteCommandItem.kt
@@ -7,6 +7,8 @@
 
 package im.vector.app.features.autocomplete.command
 
+import android.view.View
+import android.widget.ImageView
 import android.widget.TextView
 import com.airbnb.epoxy.EpoxyAttribute
 import com.airbnb.epoxy.EpoxyModelClass
@@ -15,6 +17,8 @@ import im.vector.app.core.epoxy.ClickListener
 import im.vector.app.core.epoxy.VectorEpoxyHolder
 import im.vector.app.core.epoxy.VectorEpoxyModel
 import im.vector.app.core.epoxy.onClick
+import im.vector.app.features.home.AvatarRenderer
+import org.matrix.android.sdk.api.util.MatrixItem
 
 @EpoxyModelClass
 abstract class AutocompleteCommandItem : VectorEpoxyModel<AutocompleteCommandItem.Holder>(R.layout.item_autocomplete_command) {
@@ -28,6 +32,12 @@ abstract class AutocompleteCommandItem : VectorEpoxyModel<AutocompleteCommandIte
     @EpoxyAttribute
     var description: String? = null
 
+    @EpoxyAttribute
+    var source: MatrixItem.UserItem? = null
+
+    @EpoxyAttribute
+    lateinit var avatarRenderer: AvatarRenderer
+
     @EpoxyAttribute(EpoxyAttribute.Option.DoNotHash)
     var clickListener: ClickListener? = null
 
@@ -37,9 +47,16 @@ abstract class AutocompleteCommandItem : VectorEpoxyModel<AutocompleteCommandIte
         holder.nameView.text = name
         holder.parametersView.text = parameters
         holder.descriptionView.text = description
+        source?.let {
+            holder.avatarView.visibility = View.VISIBLE
+            avatarRenderer.render(it, holder.avatarView)
+        } ?: run {
+            holder.avatarView.visibility = View.GONE
+        }
     }
 
     class Holder : VectorEpoxyHolder() {
+        val avatarView by bind<ImageView>(R.id.commandAvatar)
         val nameView by bind<TextView>(R.id.commandName)
         val parametersView by bind<TextView>(R.id.commandParameter)
         val descriptionView by bind<TextView>(R.id.commandDescription)